requirements.txt
numpy>=1.20.1
matplotlib>=3.3.4
pyserial>=3.5
scipy>=1.8.0
keyboard>=0.13.5
open3d>=0.13.0